Fix code style and comparison between different types

+++ avogadro/rendering/bsplinegeometry.cpp
@@ -47,11 +47,13 @@ Vector3f BSplineGeometry::computeCurvePoint(
   float i = 0.0f;
   auto it = points.begin();
   const auto end = points.end();
-  size_t size = points.size();
+  int size = static_cast<int>(points.size());
   // start from a lookbehind distance rather than at the beginning
   int startIndex = (size * t) - lookahead;
-  if (startIndex < 0) startIndex = 0;
-  else if (startIndex > size - 2 * lookahead) startIndex = size - 2 * lookahead;
+  if (startIndex < 0)
+    startIndex = 0;
+  else if (startIndex > size - 2 * lookahead)
+    startIndex = size - 2 * lookahead;
   float t2 = (t - startIndex / (float) size) * size / (2 * lookahead);
   for (; startIndex > 0 && it != end; --startIndex, ++it) {}
   // only read a certain number of elements from here
